requirements.txt
streamlit
pandas
psycopg2-binary
python-dotenv
opencv-python-headless
qrcode[pil]
openpyxl
